Create SUM OF THE DIGITS IS PALINDROME OR NOT
To solve this problem, we first need to compute the digit sum of the given number. Then, we need to check if this digit sum is a palindrome or not. To determine if a number is a palindrome, we can convert it to a string and compare it with its reverse.

Explanation:

We define a helper function digit_sum to compute the digit sum of a number.
We compute the digit sum of the given number n.
We convert the digit sum to a string and check if it's equal to its reverse.
If it's equal, we return "PALINDROME", otherwise, we return "NOT PALINDROME".