Create ADDITION OF TWO NUMBERS
To implement addition of two numbers we can directly use the + operator in Python. The time complexity of this operation is O(1), as it performs a constant number of arithmetic operations, regardless of the size of the numbers.

Explanation:

The addition function simply returns the sum of the two input numbers A,B
We use the + operator to perform the addition operation.