Create ODD OR EVEN
To determine whether a number is odd or even, we can simply check its remainder when divided by 2. If the remainder is 0, the number is even; otherwise, it's odd.

Explanation:

We use the modulo operator % to find the remainder when N is divided by 2.
If the remainder is 0, we return "even"; otherwise, we return "odd".