Create POWER OF POW
To find the sum of the first N even natural numbers, we can use the formula for the sum of an arithmetic series. The sum of the first 
N even natural numbers is given by:
Sum = N * (N + 1)

Explanation:

We use the formula 
Sum = N * (N + 1)
to calculate the sum of the first 
N even natural numbers.
Finally, we return the calculated sum.