Create NUMBER OF DIVISORS
The function count_divisors(N) takes an integer N as input and returns the count of divisors of N that are divisible by 3.

Inside the function:

We initialize a variable count to 0. This variable will be used to count the divisors.
We iterate through all numbers from 1 to the square root of N using a for loop.
For each i in the range:
We check if i divides N evenly (N % i == 0) and if i itself is divisible by 3 (i % 3 == 0).
If both conditions are true, it means i is a divisor of N and is divisible by 3. So, we increment the count by 1.
Additionally, if i is the square root of N (i * i == N), we need to avoid counting it twice if it is divisible by 3. So, we decrement the count by 1 in this case.
Finally, we return the value of count, which represents the count of divisors of N that are divisible by 3.

In the example usage:

We prompt the user to enter a number (N) using input().
We call the count_divisors function with the input number N and print the result.